refactor(apis): polish and converge the status, grafana, ceph, and blockdev libs
