+++ internal/apis/v1/handlers/datacenters/additionalinfo.go
@@ -3,7 +3,7 @@ package datacenters
 import (
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/base"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/nodes"
-	"github.com/bigstack-oss/cube-cos-api/internal/status"
+	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/status"
 )
 
 func getNodeLicenseStatus() base.NodeLicenseStatus {
