+++ internal/apis/v1/handlers/grafana/handlers.go
@@ -6,7 +6,7 @@ import (
 	api "github.com/bigstack-oss/cube-cos-api/internal/apis"
 	"github.com/bigstack-oss/cube-cos-api/internal/apis/v1/bodies"
 	"github.com/bigstack-oss/cube-cos-api/internal/cubecos"
-	v1 "github.com/bigstack-oss/cube-cos-api/internal/definition/v1"
+	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/grafana"
 	"github.com/gin-gonic/gin"
 )
 
@@ -61,7 +61,7 @@ func forwardHostLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch host link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genHostLink(c),
 			Enabled: true,
 		},
@@ -72,7 +72,7 @@ func forwardInstanceLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch instance link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genInstanceLink(c),
 			Enabled: true,
 		},
@@ -83,7 +83,7 @@ func forwardTopHostLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch top host link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genTopHostLink(),
 			Enabled: true,
 		},
@@ -94,7 +94,7 @@ func forwardTopInstanceLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch top instance link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genTopInstanceLink(),
 			Enabled: true,
 		},
@@ -105,7 +105,7 @@ func forwardNetworksLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch networks link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genNetworksLink(),
 			Enabled: cubecos.IsOvnSFlowEnabled(),
 		},
@@ -116,7 +116,7 @@ func forwardNetworkDevicesLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch network devices link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genNetworkDevicesLink(),
 			Enabled: cubecos.IsOvnSFlowEnabled(),
 		},
@@ -127,7 +127,7 @@ func forwardStoragesLink(c *gin.Context) {
 	bodies.SetOk(
 		c,
 		"fetch storages link successfully",
-		v1.Dashboard{
+		grafana.Dashboard{
 			Link:    genStoragesLink(),
 			Enabled: true,
 		},
