--- internal/definition/v1/aws.go
@@ -1,5 +1,0 @@
-package v1
-
-const (
-	RegionAuto = "auto"
-)
