+++ internal/definition/v1/email/email.go
@@ -6,7 +6,7 @@ import (
 	"net/smtp"
 	"strconv"
 
-	"github.com/bigstack-oss/cube-cos-api/internal/status"
+	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/status"
 )
 
 const (
