+++ internal/definition/v1/service.go
@@ -3,7 +3,7 @@ package v1
 import (
 	"fmt"
 
-	"github.com/bigstack-oss/cube-cos-api/internal/status"
+	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/status"
 )
 
 const (
